refactor: extract method placeBlock and move some code from PlayerController to BlocksController
